ecommerce_analytics/requirements.txt
# Data Processing & Analysis
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
statsmodels==0.14.0

# Deep Learning
torch==2.0.1
transformers==4.30.2
tensorflow==2.13.0
keras==2.13.1

# NLP
nltk==3.8.1
spacy==3.6.0
gensim==4.3.1

# Feature Store
feast==0.30.2

# Time Series
prophet==1.1.4
pmdarima==2.0.3

# MLOps & Experiment Tracking
mlflow==2.5.0
weights-and-biases==0.15.0
evidently==0.3.2
deepchecks==0.17.3
dvc==3.13.0

# Data Orchestration
apache-airflow==2.6.3
prefect==2.10.18

# API & Serving
fastapi==0.100.0
uvicorn==0.23.1
flask==2.3.2
streamlit==1.24.1
bentoml==1.0.22

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# Cloud & Infrastructure
boto3==1.28.0
google-cloud-storage==2.10.0
azure-storage-blob==12.17.0
kubernetes==27.2.0

# Database Connectors
pymongo==4.4.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.6

# API Clients
shopify==11.0.0
pyshopify==0.2.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0
pyyaml==6.0 